帧同步下角色移动，塔的攻击，子弹的移动和伤害

+++ UnityMobaClient/Assets/Scripts/BattleScene/Building/MainTower.cs
@@ -2,7 +2,7 @@ namespace Moba.Script.Building
 {
     public class MainTower:AbstractTower
     {
-        public override void OnLogicFrameUpdate(float deltaTime)
+        public override void OnLogicFrameUpdate(int deltaTime)
         {
             base.OnLogicFrameUpdate(deltaTime);
         }
